+++ hanghae-order/src/test/kotlin/hanghae99/plus2/team3/hanghaeorder/domain/order/small/OrderPaymentUseCaseTest.kt
@@ -43,7 +43,6 @@ class OrderPaymentUseCaseTest {
             OrderService(
                 orderRepository,
                 orderItemRepository,
-                userAccessor,
                 productsAccessor,
                 listOf(
                     PaymentTotalValidator(),
